requirements.txt
aiosqlite==0.20.0
anyio==4.4.0
APScheduler==3.10.4
certifi==2024.6.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-telegram-bot==21.3
pytz==2024.1
six==1.16.0
sniffio==1.3.1
types-pytz==2024.1.0.20240417
typing_extensions==4.12.2
tzlocal==5.2
